sdpsoft sync: better print, error handling

+++ sdpsoft_sync.sh
@@ -37,36 +37,43 @@ SERVERS=(
 
 # Print usage info if none or an invalid argument is given.
 function usage {
-    echo "usage: $PROGRAMNAME [-a, --all] [-i, --include] [--only-env-files]"
-    echo -e "      -f, --files           Only sync these files. Example: \"./sync.sh --files python2.7.14 tmux-10.7\""
+    echo "usage: $PROGRAMNAME [-a, --all] [-f, --files] [--only-env-files]"
+    echo -e "      -f, --files           Only sync these files. Example: \"./$PROGRAMNAME --files python2.7.14 tmux-*\""
     echo "      -a, --all             Sync all files"
-    echo "      --only-env-files      Only sync 'the env.sh' and 'env.csh' files"
-    echo "   requires atleast one argument"
+    echo "      --only-env-files      Only sync the 'env.sh' and 'env.csh' files"
     exit 1
 }
 
 if [ "$1" = "--only-env-files" ]; then
+    if [ ! -f "env.sh" ] || [ ! -f "env.csh" ]; then
+        echo "Error: One or more of the 'env-files' can not be found...exiting"
+        exit 1
+    fi
     for server in ${SERVERS[@]}; do
+        echo "----------------------------------------------"
         echo " SYNCING to $server"
         rsync -va --include="env.sh" --include="env.csh" --exclude="*" . $server:$SDPSOFT_REMOTE_DIR
     done
 elif [ "$1" = "-a" ] || [ "$1" = "--all" ]; then
     for server in ${SERVERS[@]}; do
+        echo "----------------------------------------------"
         echo " SYNCING to $server"
         # Sync the updated environment-files last to avoid putting the updated
         # software in limbo while the new version is syncing
-        rsync -va --exclude-from=".gitignore" --exclude="env.sh" --exclude="env.csh" . $server:$SDPSOFT_REMOTE_DIR
-        rsync -va --include="env.sh" --include="env.csh" --exclude="*" . $server:$SDPSOFT_REMOTE_DIR
+        rsync -va --delete --exclude-from=".gitignore" --exclude="env.sh" --exclude="env.csh" . $server:$SDPSOFT_REMOTE_DIR
+        rsync -va --delete --include="env.sh" --include="env.csh" --exclude="*" . $server:$SDPSOFT_REMOTE_DIR
     done
 elif [ "$1" = "-f" ] || [ "$1" = "--files" ]; then
     # Shift the $@ parameters. $2 becomes $1. To omit '--files'.
     shift 1
     for server in ${SERVERS[@]}; do
+        echo "----------------------------------------------"
         echo " SYNCING to $server"
         for file in "$@"; do
             rsync -va $file $server:$SDPSOFT_REMOTE_DIR
         done
     done
 else
+    echo "Error: Requires atleast one argument."
     usage
 fi
